Flip all interlaced fields.
It turns out the interlaced fields are in the wrong order for the
Component input as well, so always flip the interlaced field order.

AC3 via S/PDIF with Component input is working. I don't know what
change allowed for that, though.  It could be my previous testing was invalid.